refactor: 회원가입, 로그인 공통 함수 관리하는 모듈인 common.js 경로 변경
